requirements.txt
langchain>=0.3.7
langchain_core>=0.3.19
langchain_openai>=0.2.8
Pillow>=11.0.0
python-dotenv>=1.0.1
pytest>=9.0.0
pytest-xdist>=3.8.0
pytube>=15.0.0
Requests>=2.32.3
streamlit>=1.51.0
youtube_transcript_api>=1.2.3
yt-dlp>=2025.2.19
